Successfully implemented the counts pca function, which conducts pca on the counts matrix, focusing on the first 10 pcs, and creates a scatter plot where the user can select which pcs to plot on x and y axes. The axis labels in this plot are dynamic and display the % variance for each pc.
Started the counts heatmap function and tested on a small subset of data. It can create a basic heatmap with genes as rows and samples as columns. I need to figure out a way to do the following: 1.) do log10 transformation on all counts, 2.) remove default dendrograms and cluster the heatmap by control/HD status, 3.) upscale the script so that it can work efficiently on the whole dataset; currently has very long load times, and 4.) make sure the filtering functionality works